Added streamlit integration
